refactor: #90 일렉 기타에 대한 package 분리
