+++ src/main/java/com/ajou/hertz/domain/instrument/electric_guitar/dto/response/ElectricGuitarResponse.java
@@ -1,14 +1,16 @@
-package com.ajou.hertz.domain.instrument.dto.response;
+package com.ajou.hertz.domain.instrument.electric_guitar.dto.response;
 
 import java.util.List;
 
 import com.ajou.hertz.common.dto.response.AddressResponse;
-import com.ajou.hertz.domain.instrument.constant.ElectricGuitarBrand;
-import com.ajou.hertz.domain.instrument.constant.ElectricGuitarModel;
+import com.ajou.hertz.domain.instrument.dto.response.InstrumentImageResponse;
+import com.ajou.hertz.domain.instrument.dto.response.InstrumentResponse;
+import com.ajou.hertz.domain.instrument.electric_guitar.constant.ElectricGuitarBrand;
+import com.ajou.hertz.domain.instrument.electric_guitar.constant.ElectricGuitarModel;
 import com.ajou.hertz.domain.instrument.constant.GuitarColor;
 import com.ajou.hertz.domain.instrument.constant.InstrumentCategory;
 import com.ajou.hertz.domain.instrument.constant.InstrumentProgressStatus;
-import com.ajou.hertz.domain.instrument.dto.ElectricGuitarDto;
+import com.ajou.hertz.domain.instrument.electric_guitar.dto.ElectricGuitarDto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AccessLevel;
