+++ src/main/java/com/ajou/hertz/domain/instrument/electric_guitar/strategy/ElectricGuitarCreationStrategy.java
@@ -1,7 +1,8 @@
-package com.ajou.hertz.domain.instrument.strategy;
+package com.ajou.hertz.domain.instrument.electric_guitar.strategy;
 
-import com.ajou.hertz.domain.instrument.dto.request.CreateNewElectricGuitarRequest;
-import com.ajou.hertz.domain.instrument.entity.ElectricGuitar;
+import com.ajou.hertz.domain.instrument.electric_guitar.dto.request.CreateNewElectricGuitarRequest;
+import com.ajou.hertz.domain.instrument.electric_guitar.entity.ElectricGuitar;
+import com.ajou.hertz.domain.instrument.strategy.InstrumentCreationStrategy;
 import com.ajou.hertz.domain.user.entity.User;
 
 public class ElectricGuitarCreationStrategy
