+++ src/main/java/com/ajou/hertz/domain/instrument/repository/InstrumentRepositoryCustomImpl.java
@@ -5,7 +5,7 @@ import static com.ajou.hertz.domain.instrument.entity.QAmplifier.*;
 import static com.ajou.hertz.domain.instrument.entity.QAudioEquipment.*;
 import static com.ajou.hertz.domain.instrument.entity.QBassGuitar.*;
 import static com.ajou.hertz.domain.instrument.entity.QEffector.*;
-import static com.ajou.hertz.domain.instrument.entity.QElectricGuitar.*;
+import static com.ajou.hertz.domain.instrument.electric_guitar.entity.QElectricGuitar.*;
 import static com.ajou.hertz.domain.user.entity.QUser.*;
 
 import java.util.ArrayList;
@@ -25,14 +25,14 @@ import com.ajou.hertz.domain.instrument.dto.request.AmplifierFilterConditions;
 import com.ajou.hertz.domain.instrument.dto.request.AudioEquipmentFilterConditions;
 import com.ajou.hertz.domain.instrument.dto.request.BassGuitarFilterConditions;
 import com.ajou.hertz.domain.instrument.dto.request.EffectorFilterConditions;
-import com.ajou.hertz.domain.instrument.dto.request.ElectricGuitarFilterConditions;
+import com.ajou.hertz.domain.instrument.electric_guitar.dto.request.ElectricGuitarFilterConditions;
 import com.ajou.hertz.domain.instrument.dto.request.InstrumentFilterConditions;
 import com.ajou.hertz.domain.instrument.entity.AcousticAndClassicGuitar;
 import com.ajou.hertz.domain.instrument.entity.Amplifier;
 import com.ajou.hertz.domain.instrument.entity.AudioEquipment;
 import com.ajou.hertz.domain.instrument.entity.BassGuitar;
 import com.ajou.hertz.domain.instrument.entity.Effector;
-import com.ajou.hertz.domain.instrument.entity.ElectricGuitar;
+import com.ajou.hertz.domain.instrument.electric_guitar.entity.ElectricGuitar;
 import com.ajou.hertz.domain.instrument.entity.Instrument;
 import com.ajou.hertz.domain.user.entity.User;
 import com.querydsl.core.types.Order;
